[fix the bug: after cancel the timer, can't schedule the timer again]

+++ src/main/java/com/example/FileQueueService.java
@@ -1,11 +1,40 @@
 package com.example;
 
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+
 /**
  * Created by tina on 2019/2/7.
  */
 public class FileQueueService implements QueueService {
+    private Map<String, FileQueue> queues;
+
     public FileQueueService() {
-        // do nothing
+        this.queues = new ConcurrentHashMap<>();
+    }
+
+    private Map<String, FileQueue> getQueues() {
+        return this.queues;
+    }
+
+    public synchronized FileQueue getQueueByName(String queueName) {
+        Map<String, FileQueue> queues = getQueues();
+        if (queues.containsKey(queueName)) {
+            return queues.get(queueName);
+        } else {
+            return null;
+        }
+    }
+
+    public synchronized FileQueue createQueueByName(String queueName) {
+        Map<String, FileQueue> queues = getQueues();
+        if (queues.containsKey(queueName)) {
+            return queues.get(queueName);
+        } else {
+            FileQueue queue = new FileQueue(queueName);
+            queues.put(queueName, queue);
+            return queue;
+        }
     }
 
     @Override
@@ -15,13 +44,19 @@ public class FileQueueService implements QueueService {
             return false;
         }
 
+        /** Find the queue named as queueName.
+         *  If there is no specified queue, create a new queue with the queueName */
+        FileQueue fileQueue = getQueueByName(queueName);
+        if (fileQueue == null) {
+            fileQueue = createQueueByName(queueName);
+        }
+
         /** If there is nothing to be pushed, push success */
         if (messages == null || messages.length == 0) {
             return true;
         }
 
         /**  add messages one by one in the end of the queue */
-        FileQueue fileQueue = new FileQueue(queueName);
         for (String data : messages) {
             Message messageNode = new Message(data);
             fileQueue.push(messageNode);
@@ -39,7 +74,12 @@ public class FileQueueService implements QueueService {
             return null;
         }
 
-        FileQueue fileQueue = new FileQueue(queueName);
+        /** Find the queue named as queueName.
+         *  If there is no specified queue, pull fails, return null */
+        FileQueue fileQueue = getQueueByName(queueName);
+        if (fileQueue == null) {
+            return null;
+        }
 
         /** If consumer set the legal visibilityTimeout, use the set visibilityTimeout
          *  otherwise use default visibility timeout */
@@ -63,12 +103,18 @@ public class FileQueueService implements QueueService {
             return false;
         }
 
+        /** Find the queue named as queueName.
+         *  If there is no specified queue, delete fails, return false */
+        FileQueue fileQueue = getQueueByName(queueName);
+        if (fileQueue == null) {
+            return false;
+        }
+
         /** Input receiptHandle is illegal, return false */
         if (receiptHandle == null || receiptHandle.length() == 0) {
             return false;
         }
 
-        FileQueue fileQueue = new FileQueue(queueName);
         return fileQueue.delete(receiptHandle);
     }
